+++ app/models/employee_profile.rb
@@ -25,8 +25,8 @@ class EmployeeProfile < ApplicationRecord
   enum marital_status: { single: 1, married: 4, widower: 8, divorced: 10 }
 
   def value
-    card_type = GetCardType.find(self.position.card_type_id, self.department.company.registration_number)
-    card_type ? card_type.start_points : 0
+    card_type = GetCardType.find(position.card_type_id, department.company.registration_number)
+    card_type ? card_type.start_points.to_f : 0
   end
 
   private
